+++ .github/CHANGELOG.md
@@ -15,6 +15,7 @@
 1. [ATSU] Add support for BeyondATC and SayIntentions AI as METAR/ATIS sources - @saschl
 1. [A32NX/TCAS] Improve PFD Resolution Advisory graphics, introduce TCAS bus - @lukecologne (luke)
 1. [PFD] Reduced size of white FMA messages on PFD - @BravoMike99 (bruno_pt99)
+1. [A380X/MFD] Show approach QNH and TEMP as mandatory only when closer than 180 NM to destination or when predictions are unavailable - @matze-tech (matze2346)
 
 ## 2024.1.0